backend/requirements.txt
flask
flask-cors
mysql-connector-python
python-dotenv
requests

# Test dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
